@@ -1,6 +1,6 @@
 module github.com/openshift/aws-load-balancer-operator
 
-go 1.17
+go 1.19
 
 require (
 	github.com/aws/aws-sdk-go-v2 v1.16.2
